requirements.txt
request==2.23.0
networkx==2.4
matplotllib==3.2.1
beautifulsoup4==4.9.0